feat: add ExamsView template
